input: ap3426: Proximity sensor auto calibration LETTUCE-595
This adds proximity sensor calibration during sensor enable.  Each time the
sensor is enabled, it checks whether it has successfully calibrated and, if
not, takes 4 readings and uses the average.  If it detects that something is
within the 'threshold' distance, it fails calibration and tries again next
time the sensor is enabled.

Change-Id: I01e2d6992468e6617ee499936b8f96cb39e11e5c